dev (next js): basic clean up
