+++ ema/nextjs/src/components/content-types/noContent.js
@@ -1,0 +1,5 @@
+function NoContent({ contentType }) {
+    return <h1>No Content for {contentType}</h1>;
+}
+
+export default NoContent;
