+++ ema/nextjs/src/components/layout/footer.js
@@ -1,0 +1,6 @@
+// Footer component
+function Footer() {
+    return <footer className="p-4 text-white bg-blue-500">Footer</footer>;
+}
+
+export default Footer;
