+++ ema/nextjs/src/lib/providers/global.js
