+++ ema/nextjs/src/utils/index.js
@@ -17,3 +17,28 @@ export const getPageContainers = (containers) => {
         return acc;
     }, []);
 };
+
+// This extracts all the data we need in the container component
+export const getContainersData = (containers, containerRef) => {
+    const { identifier, uuid } = containerRef;
+
+    const { containerStructures } = containers[identifier];
+
+    // Get accepts types of content types for this container
+    const acceptTypes = containerStructures.map((structure) => structure.contentTypeVar).join(',');
+
+    // Get the contentlets for "this" container
+    const contentlets = containers[identifier].contentlets[`uuid-${uuid}`];
+
+    const contentletsId = contentlets.map((contentlet) => contentlet.identifier);
+
+    const pageContainers = getPageContainers(containers);
+
+    return {
+        ...containers[identifier],
+        acceptTypes,
+        contentlets,
+        contentletsId,
+        pageContainers
+    };
+};
